+++ src/backend/services/basket.api/database/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "database",
+    srcs = ["db.go"],
+    importpath = "github.com/example/project/database",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_gomodule_redigo//redis"],
+)
